+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/FeedbackQueryTimeMode.cs
@@ -8,11 +8,16 @@ namespace Azure.AI.MetricsAdvisor
     /// <summary>
     /// Specifies which time property of a <see cref="MetricFeedback"/> will be used to filter results
     /// in the <see cref="MetricsAdvisorClient.GetAllFeedback"/> and the <see cref="MetricsAdvisorClient.GetAllFeedbackAsync"/>
-    /// operations.
+    /// operations. Defaults to <see cref="None"/>.
     /// </summary>
     [CodeGenModel("FeedbackQueryTimeMode")]
     public enum FeedbackQueryTimeMode
     {
+        /// <summary>
+        /// No time filter is applied.
+        /// </summary>
+        None,
+
         /// <summary>
         /// Filters feedbacks by its metric's timestamps.
         /// </summary>